requirements.txt
#torchvision==0.7.0+cu101
cityscapesScripts==2.2.0
labelme==3.16.7
pycocotools==2.0
scikit_image==0.18.3
opencv_python==4.5.3.56
matplotlib==3.2.2
requests==2.21.0
numpy==1.19.5
h5py==3.1.0
scipy==1.7.1
tqdm==4.62.2
yacs==0.1.8
#torch==1.6.0+cu101
Pillow==8.4.0
